core_llm: cast Qwen3-Omni's processed inputs to the model's dtype (fixes float/BFloat16 mismatch)
QwenOmniModel.chat() only moved apply_chat_template()'s output to the
model's device, not its dtype -- GemmaAudioModel already did both
(.to(device, dtype=model.dtype)), just missed here. The processor's
audio features come out float32 by default; the model's weights load
as bf16, so the mismatch surfaced as "Input type (float) and bias type
(c10::BFloat16) should be the same" during generate(). BatchFeature.to()
only casts floating-point tensors when given a dtype, leaving integer
tensors like input_ids untouched, so this is safe to apply broadly
(same as the already-working Gemma path).

+++ Core_LLM/deployment/model.py
@@ -196,7 +196,7 @@ class QwenOmniModel(BaseLLM):
         inputs = self._processor.apply_chat_template(
             converted, add_generation_prompt=True,
             tokenize=True, return_dict=True, return_tensors="pt", padding=True,
-        ).to(self._model.device)
+        ).to(self._model.device, dtype=self._model.dtype)
         # Slicing off input_len (unlike the reference docs snippet, which decodes
         # the full sequence) so the reply doesn't echo the prompt back -- same
         # reasoning as GemmaAudioModel above.
